refactor(codex): isolate strict hook source loading
