+++ src/codex_plugin_scanner/guard/adapters/codex.py
@@ -32,10 +32,6 @@ from ..codex_hook_integrity import (
     write_hook_manifest,
 )
 from ..codex_hook_inventory import (
-    CODEX_HOOK_INVENTORY_SOURCE_CHANGED,
-    CODEX_HOOK_INVENTORY_SOURCE_DUPLICATE,
-    CODEX_HOOK_INVENTORY_SOURCE_MALFORMED,
-    CODEX_HOOK_INVENTORY_SOURCE_UNREADABLE,
     CODEX_HOOK_INVENTORY_UNMANAGED_EXECUTABLE,
     CodexHookInventory,
     enumerate_codex_hooks,
@@ -58,6 +54,11 @@ from ..codex_hook_registration import (
 from ..codex_hook_registration import (
     remove_manifest_bound_hook_events as _remove_manifest_bound_hook_events,
 )
+from ..codex_hook_sources import (
+    require_hook_inventory_sources_unchanged as _require_hook_inventory_sources_unchanged,
+)
+from ..codex_hook_sources import strict_json_object as _strict_json_object
+from ..codex_hook_sources import strict_toml_object as _strict_toml_object
 from ..config import MAX_APPROVAL_WAIT_TIMEOUT_SECONDS, load_guard_config, resolve_guard_home
 from ..launcher import merge_guard_launcher_env
 from ..models import GuardArtifact, HarnessDetection
@@ -218,107 +219,6 @@ def _json_object(path: Path) -> dict[str, object]:
     except (OSError, json.JSONDecodeError):
         return {}
     return payload if isinstance(payload, dict) else {}
-
-
-def _strict_json_object(path: Path, *, label: str) -> dict[str, object]:
-    if path.is_symlink() or (path.exists() and not path.is_file()):
-        raise RuntimeError(
-            f"{CODEX_HOOK_INVENTORY_SOURCE_UNREADABLE}: Guard refused to overwrite non-file {label} at {path}. "
-            "Replace it with a readable regular file before retrying install."
-        )
-    if not path.is_file():
-        return {}
-    try:
-        payload = json.loads(path.read_text(encoding="utf-8"), object_pairs_hook=_unique_json_object)
-    except _DuplicateJsonKeyError as exc:
-        raise RuntimeError(
-            f"{CODEX_HOOK_INVENTORY_SOURCE_DUPLICATE}: Guard found duplicate key {exc.key!r} in {label} at "
-            f"{path}. Remove the duplicate before retrying install."
-        ) from exc
-    except OSError as exc:
-        raise RuntimeError(
-            f"{CODEX_HOOK_INVENTORY_SOURCE_UNREADABLE}: Guard could not read {label} at {path}. Repair file "
-            "permissions before retrying install."
-        ) from exc
-    except (UnicodeDecodeError, json.JSONDecodeError) as exc:
-        raise RuntimeError(
-            f"{CODEX_HOOK_INVENTORY_SOURCE_MALFORMED}: Guard refused to overwrite unreadable {label} at {path} "
-            "because its JSON is malformed. Repair the JSON before retrying install."
-        ) from exc
-    if not isinstance(payload, dict):
-        raise RuntimeError(
-            f"{CODEX_HOOK_INVENTORY_SOURCE_MALFORMED}: Guard refused to overwrite non-object {label} at {path}. "
-            "Replace it with a JSON object before retrying install."
-        )
-    return payload
-
-
-class _DuplicateJsonKeyError(ValueError):
-    key: str
-
-    def __init__(self, key: str) -> None:
-        super().__init__(key)
-        self.key = key
-
-
-def _unique_json_object(pairs: list[tuple[str, object]]) -> dict[str, object]:
-    payload: dict[str, object] = {}
-    for key, value in pairs:
-        if key in payload:
-            raise _DuplicateJsonKeyError(key)
-        payload[key] = value
-    return payload
-
-
-def _strict_toml_object(path: Path, *, label: str) -> dict[str, object]:
-    if path.is_symlink() or (path.exists() and not path.is_file()):
-        raise RuntimeError(
-            f"{CODEX_HOOK_INVENTORY_SOURCE_UNREADABLE}: Guard refused to overwrite non-file {label} at {path}. "
-            "Replace it with a readable regular file before retrying install."
-        )
-    if not path.is_file():
-        return {}
-    try:
-        with path.open("rb") as handle:
-            payload = tomllib.load(handle)
-    except OSError as exc:
-        raise RuntimeError(
-            f"{CODEX_HOOK_INVENTORY_SOURCE_UNREADABLE}: Guard could not read {label} at {path}. Repair file "
-            "permissions before retrying install."
-        ) from exc
-    except tomllib.TOMLDecodeError as exc:
-        reason = (
-            CODEX_HOOK_INVENTORY_SOURCE_DUPLICATE
-            if "overwrite" in str(exc).lower()
-            else CODEX_HOOK_INVENTORY_SOURCE_MALFORMED
-        )
-        raise RuntimeError(
-            f"{reason}: Guard could not parse {label} at {path}. Repair the TOML before retrying install."
-        ) from exc
-    if not isinstance(payload, dict):
-        raise RuntimeError(
-            f"{CODEX_HOOK_INVENTORY_SOURCE_MALFORMED}: Guard refused to overwrite non-object {label} at {path}."
-        )
-    return payload
-
-
-def _require_hook_inventory_sources_unchanged(
-    *,
-    config_payloads: Mapping[Path, dict[str, object]],
-    hook_payloads: Mapping[Path, dict[str, object]],
-) -> None:
-    for path, expected in config_payloads.items():
-        if _strict_toml_object(path, label="Codex config file") != expected:
-            raise RuntimeError(
-                f"{CODEX_HOOK_INVENTORY_SOURCE_CHANGED}: Codex config changed after pre-activation inventory at "
-                f"{path}. Retry install after configuration writes have stopped."
-            )
-    for path, expected in hook_payloads.items():
-        if _strict_json_object(path, label="Codex hooks file") != expected:
-            raise RuntimeError(
-                f"{CODEX_HOOK_INVENTORY_SOURCE_CHANGED}: Codex hooks changed after pre-activation inventory at "
-                f"{path}. Retry install after configuration writes have stopped."
-            )
 
 
 def _local_hook_command_parts_for_home_mode(
